refactor(yomu): tools/format.rs から markdown renderer を抽出 (#137)
RC-009 (#137) の file 分割。tools/format.rs 520→315 行に縮小し、#255 の
file-size gate allowlist から除去する。

- markdown rendering (search/impact 結果) → src/tools/format/render.rs (296 行)
- JSON envelope formatter (format_results_json 等) + coverage helper は
  format.rs に残す

分割方向の判断: JSON を残すことで (a) format/tests.rs の JSON 内部テスト
(JsonChunk/JsonResponse) を `use super::*` のまま温存、(b) json↔text 共有の
coverage_if_partial を cross-module 呼び出しにしない。

公開 API (EnrichmentContext / format_impact_all / format_impact_results /
format_results_grouped) は format.rs から `pub(super) use render::{}` で
再エクスポート → 兄弟 (search.rs / impact.rs) の `use super::format::{}` 不変。
render 項目は再エクスポート制約 (E0364) 上 pub(crate) だが、`mod render` が
private のため実効到達は tools 内に限定 (理由は render.rs に明記)。内部 render
helper (dependents/file_group/imports/siblings/semantic_section) は private。

semantic_related レンダリング経路 (分割前から未カバー、新ファイルで diff-cover
の追加行となる) の直接ユニットテスト 2 件 (T-706/T-707) を render.rs に追加。

/audit 反映: pub(crate) の E0364 理由をコメント化、EnrichmentContext に
#[derive(Debug)] (LANG.md)、module doc を coverage helper 込みに更新、
fixture の chunk_id を None に。

挙動保存: 720 tests pass (4 skipped、うち新規 2)。diff-cover 99.6%。

Refs #137
