+++ src/tools/format.rs
@@ -1,19 +1,22 @@
-//! JSON envelope formatters for success-path routes.
+//! JSON envelope formatters for success-path routes, plus shared
+//! index-coverage helpers (used by both JSON and text output paths).
 //!
 //! `degraded` and `notes` fields are always serialized (no `skip_serializing_if`)
 //! so consumers can read `obj.degraded` without an existence guard
 //! (OUTCOME.md Behavior #4, spec FR-001 / BR-002).
 
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 
 use serde::Serialize;
 
+use crate::injection_check::InjectionCheck;
 use crate::{indexer, storage};
 
-pub(super) struct EnrichmentContext {
-    pub imports: HashMap<String, String>,
-    pub siblings: HashMap<String, Vec<storage::SiblingInfo>>,
-}
+mod render;
+
+pub(super) use render::{
+    EnrichmentContext, format_impact_all, format_impact_results, format_results_grouped,
+};
 
 pub(super) fn format_coverage(stats: &storage::IndexStatus) -> String {
     format!(
@@ -44,214 +47,6 @@ pub(super) fn format_coverage_note(stats: &storage::IndexStatus) -> Option<Strin
         format!("\n\nEmbedding coverage: {cov}. Run `yomu index` again to finish embedding.")
     })
 }
-
-pub(super) fn format_dependents_by_depth(
-    output: &mut String,
-    dependents: &[storage::Dependent],
-    heading_prefix: &str,
-) {
-    debug_assert!(
-        dependents.windows(2).all(|w| w[0].depth <= w[1].depth),
-        "dependents must be sorted by depth"
-    );
-    let mut current_depth = 0;
-    for dep in dependents {
-        if dep.depth != current_depth {
-            current_depth = dep.depth;
-            output.push_str(&format!("{heading_prefix} Depth {}\n", current_depth));
-        }
-        output.push_str(&format!("- {}\n", dep.file_path));
-    }
-}
-
-fn format_semantic_section(output: &mut String, semantic_related: &[storage::SearchResult]) {
-    output.push_str("\n### Semantic related (embedding search)\n");
-    for r in semantic_related {
-        output.push_str(&format!(
-            "- {} (similarity: {:.2})\n",
-            r.chunk.file_path, r.score
-        ));
-    }
-}
-
-pub(super) fn format_impact_all(
-    target: &str,
-    dependents: &[storage::Dependent],
-    semantic_related: &[storage::SearchResult],
-) -> String {
-    let mut output = format!("## Impact analysis: `{}`\n\n", target);
-
-    if semantic_related.is_empty() {
-        format_dependents_by_depth(&mut output, dependents, "###");
-    } else {
-        output.push_str("### Structural dependents (import graph)\n");
-        format_dependents_by_depth(&mut output, dependents, "####");
-        format_semantic_section(&mut output, semantic_related);
-    }
-
-    output.push_str(&format!(
-        "\nTotal: {} dependent file(s)\n",
-        dependents.len()
-    ));
-    output
-}
-
-pub(super) fn format_impact_results(
-    target: &str,
-    symbol_refs: &[String],
-    all_dependents: &[storage::Dependent],
-    semantic_related: &[storage::SearchResult],
-) -> String {
-    let mut output = format!("## Impact analysis: `{}`\n\n", target);
-
-    if !symbol_refs.is_empty() {
-        output.push_str("### Direct symbol references\n");
-        for f in symbol_refs {
-            output.push_str(&format!("- {}\n", f));
-        }
-    }
-
-    output.push_str("\n### All transitive dependents\n");
-    format_dependents_by_depth(&mut output, all_dependents, "####");
-
-    if !semantic_related.is_empty() {
-        format_semantic_section(&mut output, semantic_related);
-    }
-
-    output.push_str(&format!(
-        "\nTotal: {} dependent file(s)\n",
-        all_dependents.len()
-    ));
-    output
-}
-
-fn format_imports_line(imports: &HashMap<String, String>, file_path: &str) -> Option<String> {
-    let imports_text = imports.get(file_path)?;
-    let items: Vec<&str> = imports_text.split('\n').filter(|s| !s.is_empty()).collect();
-    if items.is_empty() {
-        return None;
-    }
-    Some(format!("Imports: {}\n", items.join(", ")))
-}
-
-fn format_siblings_line(
-    siblings_map: &HashMap<String, Vec<storage::SiblingInfo>>,
-    file_path: &str,
-    result_ranges: &HashSet<(u32, u32)>,
-) -> Option<String> {
-    let siblings = siblings_map.get(file_path)?;
-    let filtered: Vec<String> = siblings
-        .iter()
-        .filter(|s| !result_ranges.contains(&(s.start_line, s.end_line)))
-        .map(|s| {
-            let name = s.name.as_deref().unwrap_or("(unnamed)");
-            format!("{} [{}]", name, s.chunk_type.as_str())
-        })
-        .collect();
-    if filtered.is_empty() {
-        return None;
-    }
-    Some(format!("Siblings: {}\n", filtered.join(", ")))
-}
-
-fn format_file_group(
-    output: &mut String,
-    file_path: &str,
-    chunks: &[(usize, &storage::SearchResult)],
-    ctx: &EnrichmentContext,
-    parent_chunks: &HashMap<i64, storage::Chunk>,
-    result_chunk_ids: &HashSet<i64>,
-) {
-    output.push_str(&format!("## {}\n", file_path));
-
-    if let Some(line) = format_imports_line(&ctx.imports, file_path) {
-        output.push_str(&line);
-    }
-
-    let result_ranges: HashSet<(u32, u32)> = chunks
-        .iter()
-        .map(|(_, r)| (r.chunk.start_line, r.chunk.end_line))
-        .collect();
-    if let Some(line) = format_siblings_line(&ctx.siblings, file_path, &result_ranges) {
-        output.push_str(&line);
-    }
-
-    output.push('\n');
-
-    for (rank, result) in chunks {
-        let chunk = &result.chunk;
-        let name = chunk.name.as_deref().unwrap_or("(unnamed)");
-        output.push_str(&format!(
-            "{}. {} [{}] — {}:{} (similarity: {:.2})\n",
-            rank + 1,
-            name,
-            chunk.chunk_type.as_str(),
-            chunk.start_line,
-            chunk.end_line,
-            result.score,
-        ));
-        output.push_str(&chunk.content);
-        output.push_str("\n\n");
-
-        if let Some(parent_id) = chunk.parent_chunk_id
-            && !result_chunk_ids.contains(&parent_id)
-            && let Some(parent) = parent_chunks.get(&parent_id)
-        {
-            let parent_name = parent.name.as_deref().unwrap_or("(unnamed)");
-            output.push_str(&format!(
-                "  Parent context: {} [{}] — {}:{}\n",
-                parent_name,
-                parent.chunk_type.as_str(),
-                parent.start_line,
-                parent.end_line,
-            ));
-            output.push_str(&parent.content);
-            output.push_str("\n\n");
-        }
-    }
-}
-
-pub(super) fn format_results_grouped(
-    results: &[storage::SearchResult],
-    ctx: &EnrichmentContext,
-    parent_chunks: &HashMap<i64, storage::Chunk>,
-) -> String {
-    let mut groups: HashMap<&str, Vec<(usize, &storage::SearchResult)>> = HashMap::new();
-    for (i, result) in results.iter().enumerate() {
-        groups
-            .entry(&result.chunk.file_path)
-            .or_default()
-            .push((i, result));
-    }
-
-    let mut sorted: Vec<_> = groups.into_iter().collect();
-    sorted.sort_by(|a, b| {
-        let best = |items: &[(usize, &storage::SearchResult)]| {
-            items
-                .iter()
-                .map(|(_, r)| r.score)
-                .fold(f32::NEG_INFINITY, f32::max)
-        };
-        best(&b.1).total_cmp(&best(&a.1))
-    });
-
-    let mut output = String::new();
-    let result_chunk_ids: HashSet<i64> = results.iter().filter_map(|r| r.chunk_id).collect();
-
-    for (file_path, chunks) in &sorted {
-        format_file_group(
-            &mut output,
-            file_path,
-            chunks,
-            ctx,
-            parent_chunks,
-            &result_chunk_ids,
-        );
-    }
-    output
-}
-
-use crate::injection_check::InjectionCheck;
 
 #[derive(Serialize)]
 struct JsonChunk<'a> {
